+++ src/cli/pullValuesFromCommand.js
