--- src/constants/valid-flags.js
@@ -1,3 +1,0 @@
-const validFlags = ['rhyme', 'nearRhyme', 'synonym']
-
-module.exports = validFlags
