+++ src/rhymes.js
@@ -1,0 +1,13 @@
+function rhyme(rhymeInfo) {
+  return {
+    getSyllables: () => rhymeInfo.numSyllables,
+    word: () => rhymeInfo.word,
+    score: () => rhymeInfo.score,
+  }
+}
+
+function rhymes(rhymesData) {
+  return {
+    getRhymes: () => rhymesData
+  }
+}
